+++ packages/uikit/src/components/index.ts
@@ -56,3 +56,5 @@ export * from "./Waterfall";
 export * from "./WorkflowGraph";
 
 export * from "./TabRow";
+
+export * from "./ListSearchInput";